+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ### master
 - bugfix: with vim 'session' strategy, if the session file does not exist - make
-  sure wim does not contain `-S` flag
+  sure vim does not contain `-S` flag
+- enable restoring programs with arguments (e.g. "rails console") and also
+  processes that contain program name
 
 ### v0.1.0, 2014-08-28
 - refactor checking if saved tmux session exists